Conflicts:
	Applications/Canvas/Main.cpp
	Applications/LaunchStack/AssetClient.cpp
	Applications/LaunchStack/AssetClient.hpp
	Applications/MultiPluginDriver/Main.cpp
	Applications/TimeLine/Main.cpp
	Examples/HelloImages/Main.cpp
	Examples/ImageGrid/ImageGrid.cpp
	Examples/MultiAnimate/AnimatorWidget.hpp
	Examples/PluginLoader/Main.cpp
	Examples/ThreadedCanvas/Main.cpp
	Examples/TouchBrowser/Main.cpp
	Examples/TouchBrowser2/Main.cpp
	Fluffy/LayoutEngine.cpp
	Fluffy/LayoutEngine.hpp
	Fluffy/StyleSheet.cpp
	Fluffy/StyleSheet.hpp
	Fluffy/Styleable.cpp
	Fluffy/Styleable.hpp
	Fluffy/ValueVector.hpp
	MultiBrowser/BrowserWidget.cpp
	MultiBrowser2/BrowserFrame.cpp
	MultiBrowser2/BrowserFrame.hpp
	MultiBrowser2/BrowserWidget.cpp
	MultiTouch/HWControl.cpp
	MultiTouch/HandShakeServer.cpp
	MultiTouch/HardwareDetector.cpp
	MultiTouch/ImageProcessingPlugin.cpp
	MultiTouch/IndividualCamera.cpp
	MultiTouch/MainServer.cpp
	MultiTouch/Merge.cpp
	MultiTouch/MergedCamera.cpp
	MultiTouch/MultiTouch.cpp
	MultiTouch/OscReceivedElements.h
	MultiTouch/ServerThread.cpp
	MultiWidgets/AfterIdleOperator.cpp
	MultiWidgets/AfterIdleOperator.hpp
	MultiWidgets/AnimatorImpl.hpp
	MultiWidgets/Animators.cpp
	MultiWidgets/Animators.hpp
	MultiWidgets/Application.cpp
	MultiWidgets/Application.hpp
	MultiWidgets/BigImageMovie.cpp
	MultiWidgets/Calibration.cpp
	MultiWidgets/GameWidget.hpp
	MultiWidgets/ImageMovie.cpp
	MultiWidgets/ImageMovie.hpp
	MultiWidgets/ImageSkinWidget.cpp
	MultiWidgets/ImageWidget.cpp
	MultiWidgets/ItemFlow.cpp
	MultiWidgets/ItemFlow.hpp
	MultiWidgets/Keyboard.cpp
	MultiWidgets/Keyboard.hpp
	MultiWidgets/KeyboardListener.hpp
	MultiWidgets/LimitScaleByParentOperator.cpp
	MultiWidgets/ListWidget.cpp
	MultiWidgets/ListWidget.hpp
	MultiWidgets/MessageSendOperator.cpp
	MultiWidgets/MessageSendOperator.hpp
	MultiWidgets/OnInteractionOperator.cpp
	MultiWidgets/Operators.cpp
	MultiWidgets/Plugins.cpp
	MultiWidgets/RotateTowardsHandsOperator.cpp
	MultiWidgets/ScrollList.cpp
	MultiWidgets/SimpleSDLApplication.cpp
	MultiWidgets/SimpleSDLApplication.hpp
	MultiWidgets/TextBox.cpp
	MultiWidgets/TextEdit.cpp
	MultiWidgets/TextEdit.hpp
	MultiWidgets/VideoControlWidget.cpp
	MultiWidgets/VideoWidget.cpp
	MultiWidgets/ViewWidget.cpp
	MultiWidgets/Widget.cpp
	MultiWidgets/Widget.hpp
	MultiWidgets/WidgetRemoveOperator.cpp
	MultiWidgets/WidgetRemoveOperator.hpp
	Plugins/Book3D/BookWidget.cpp
	Plugins/Browsers/Browsers.cpp
	Plugins/Effects/Strangers.cpp
	Plugins/Effects/Strangers.hpp
	Plugins/Flow/Flow.cpp
	Plugins/HandMenu/BookCover.cpp
	Plugins/HandMenu/HandMenu.cpp
	Plugins/HandMenu/MenuButton.cpp
	Plugins/HandMenu/VideoPlayer.cpp
	Plugins/LaunchList/LaunchList.cpp
	Plugins/LaunchList/LaunchList.hpp
	Plugins/LaunchStack/LaunchStack.cpp
	Plugins/LaunchStack/LaunchStack.hpp
	Plugins/LightBeings/Beings.cpp
	Plugins/MarkerLauncher/MarkerLauncher.cpp
	Plugins/MarkerLauncherXML/MarkerLauncherXML.cpp
	Plugins/MindMap/HandleBox.cpp
	Plugins/Present/LoadSlideTask.cpp
	Plugins/Present/LoadSlideTask.hpp
	Plugins/Present/PollDirectoryTask.cpp
	Plugins/Present/PollDirectoryTask.hpp
	Plugins/Present/PresentWidget.cpp
	Plugins/Present/PresentWidget.hpp
	Plugins/TimeLinePlugin/Canvas.hpp
	Plugins/TimeLinePlugin/Circles.cpp
	Plugins/TimeLinePlugin/Container.cpp
	Plugins/TimeLinePlugin/EnlargerOperator.hpp
	Plugins/TimeLinePlugin/ItemLoader.cpp
	Plugins/TimeLinePlugin/KeyPointList.cpp
	Plugins/TimeLinePlugin/MainWidget.cpp
	Plugins/TwitterBall/Ecosystem.cpp
	Plugins/TwitterBall/Item.cpp
	Plugins/TwitterBall/SphereWidget.cpp
	Scripts/linux/install-ubuntu-1004-deps
	multitude/Luminous/CPUMipmapStore.cpp
	multitude/Luminous/CPUMipmaps.cpp
	multitude/Luminous/CodecRegistry.cpp
	multitude/Luminous/Image.cpp
	multitude/Luminous/Luminous.cpp
	multitude/Luminous/PixelFormat.cpp
	multitude/Luminous/RenderTarget.cpp
	multitude/Poetic/CPUWrapperFont.hpp
	multitude/Poetic/Face.hpp
	multitude/Poetic/FontManager.cpp
	multitude/Poetic/Utils.cpp
	multitude/Poetic/Utils.hpp
	multitude/Radiant/CSVDocument.cpp
	multitude/Radiant/CameraDriver.cpp
	multitude/Radiant/DateTime.cpp
	multitude/Radiant/FileUtils.cpp
	multitude/Radiant/Radiant.pro
	multitude/Radiant/Trace.cpp
	multitude/Resonant/AudioFileHandler.cpp
	multitude/Resonant/ModulePanner.cpp
	multitude/Resonant/ModuleSamplePlayer.cpp
	multitude/Screenplay/VideoFFMPEG.cpp
	multitude/Valuable/HasValues.cpp
	multitude/Valuable/Serializer.hpp
	multitude/Valuable/ValueContainer.hpp
	multitude/Valuable/ValueString.hpp
	multitude/Valuable/ValueStringImpl.hpp
	multitude/VideoDisplay/SubTitles.cpp
	multitude/VideoDisplay/VideoIn.cpp
	multitude/VideoDisplay/VideoInFFMPEG.cpp
	multitude/multitude.pri

+++ multitude/Applications/MoviePlayer/VideoWindow.cpp
@@ -105,7 +105,7 @@ void VideoWindow::randomOperation()
   };
 
   // Select movie object to stress:
-  int index = m_rand.randN24(m_movies.size());
+  int index = m_rand.randN24((uint32_t) m_movies.size());
 
   iterator it = m_movies.begin();
 
@@ -253,7 +253,7 @@ void VideoWindow::paintGL()
 
   ALL_MOVIES(update());
 
-  int n = m_movies.size();
+  int n = (int) m_movies.size();
 
   int rows = (int) ceilf(sqrtf(n));
   int cols = 1;
